Remove bib index rules from raw marc template
Remove parsing and use of bib index rules because its use as a
"secondary" check for mapping var fields to display labels has fallen
out of favor in our template; We now draw exclusively from webpub.def.

Also corrects a test broken with the move to check fieldTag (in addition
to marcTag and indicators).