Add fixtured test of annotated-marc-endpoint
Adds a test of the annotatedMarc resources controller method to perform
a deeper test of the controller method's ability to fetch a bib record
from the platform api and resolve an annotated-marc serialization.
(Sibling tests hit the serializer directly, passing bib data inline.)
This also confirms the now-centralized nypl-data-api-client is working
and can be tested using fixtures in the same way we're testing patron
service calls.