SCC-2299 Update reserved character escaping
This resolves a bug that was preventing strings that contained characters on ElasticSearch's reserved character list from being queried.

The root of the issue was that these characters were being escaped twice (e.g. `\\\\?`) which actually rendered in ES as `\?`. Paring back this to a single escape resolves the problem. Paradoxically the nature of this bug made it appear as though characters were not being escaped at all.

Additionally this adds some other improvements to the reserved character escaping method:

- This adds all of the reserved characters from the list in the ES documentation: https://www.elastic.co/guide/en/elasticsearch/reference/5.3/query-dsl-query-string-query.html#_reserved_characters
- This adds a special case as noted in that documentation that `<` and `>` must be removed entirely
- This also adds a comment to note that forward slashes are also treated as a special case

Note: The one character not included in the escape list is `"` as quoting strings for exact matches is common enough behavior that it should be left as-is.