Move logger.info for 404s into error handler
Moves the call to logger.info, which was previously in the controller
into the common error handler in the router. Because the router isn't
tested by the existing "resources*" unit tests, I had to remove some of
the assertions on logging. (I tried to add router-level unit tests, but
an into too many issues for this small change, so abandoned it. We'll
need to address better coverage with a future refactor.)