Fix order of customer code lookup and EDD determination
Fix order of 1) the live query on the SCSB API to look up item customer
codes (performed when bib is retrieved with a specific item id in the
path) and 2) the process to update the `eddRequestable` property based
on the retrieved customer code. Previously, the SCSB API query was being
performed _after_ updating `eddRequestable`. Now, the customer code
lookup is performed in parallel to building the barcode-availability
mapping (another SCSB API call that is independent of the customer code
lookup)