Hide delivery loc.s when item not requestable
Previously, item requestability of NYPL offsite items was
solely determined by ReCAP status and item holdingLocation
requestability (via NYPL-Core). See:
https://github.com/NYPL/discovery-api/blob/fecd95487027c7beb12234188d98de3a781e5805/lib/availability_resolver.js#L116

With separate request buttons, we now support three different notions of
requestability: eddRequestable, physRequestable, and specRequestable,
with `requestable` being true when any of the others are true. The
trouble is, the previous calculation of `requestable` did not take into
account EDD requestability (leading to many occasions where we did not
present a Request button for items that were EDD requestable but not
physically requestable). Now that that bug is patched, we allow patrons
to visit the Hold Request page on RC for items that are _only_ EDD
requestable in ReCAP. This is good, but uncovers another issue:

The delivery locations endpoint simply presents the valid
delivery locations based on the item's customer code regardless of the
item's phys requestability. Thus, an item in rccd8 (not requestable)
with customer code NS (deliverable to at least one other location),
appears to have at least one delivery location even though it's not
physRequestable

This fix adds a check to ensure, before providing delivery location
options, that the item is available for phys request (i.e. its status
and holdingLocation are requestable)