Fix bug parsing merge_checkin_cards param
Fix bug parsing merge_checkin_cards param due to the router function
doing some param parsing before passing it off to the controller, which
also does param parsing. This caused the boolean merge_checkin_cards
param to be converted into a boolean, which caused later param parsing
to fail to parse it, forcing it to default. Param parsing should only
be done in the controller.