Fix bug where physRequestable sometimes missing
Fix issue where items that have no deliveryLocations calculate
physRequestable=undefined, causing the property to not be added to the
API response. The fix is to cast `deliveryLocations` as a boolean, to
ensure that when `deliveryLocations` is falsey, the result of the
expression does not default to the actual value of `deliveryLocations`
which is `undefined`.