Fix EDD status check to exclude status 'o'
Fixes an issue where we are currently enabling Request Scan buttons on on-site
materials that have item status 'Use in library' (o). Sierra rejects all holds
on all items with that status. I think we only ever thought that we might want
to enable on-site EDD requests on such items because that status was part of
the original onsite-edd criteria doc. That doc was updated to remove the
status, but we retain policies in discovery-api and DFE that allow holds on
items with that status.