Fix issue when recordType not recognized in NYPL-Core
Currently, if the indexed recordTypeId is not recognized in NYPL-core,
serialization fails, affecting search and bib endpoints. Additionally,
it's possible for the recordType aggregation to be served without a label.
This change ensures that invalid recordTypes are removed from both bib
serialization and aggs. In general we should strive to keep indexed recordTypes
aligned with NYPL-Core; This update handles the case where they're not.